test: Centralize beforeEach/afterEach logic of integration tests
To avoid some duplication.

Change-type: patch
