+++ test/integration/ui/helpers.js
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2018 resin.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+const puppeteer = require('puppeteer')
+const helpers = require('../server/helpers')
+
+exports.browser = {
+	beforeEach: async (test) => {
+		await helpers.server.beforeEach(test)
+
+		const options = {
+			headless: !process.env.PUPPETEER_VISUAL_MODE,
+			args: [
+				'--window-size=1366,768',
+
+				// Set extra flags so puppeteer runs on docker
+				'--no-sandbox',
+				'--disable-setuid-sandbox'
+			]
+		}
+
+		test.context.browser = await puppeteer.launch(options)
+		test.context.page = await test.context.browser.newPage()
+		test.context.page.setViewport({
+			width: 1366,
+			height: 768
+		})
+	},
+
+	afterEach: async (test) => {
+		await test.context.browser.close()
+		await helpers.server.afterEach(test)
+	}
+}
